+++ mysql/schema.sql
@@ -55,6 +55,37 @@ CREATE UNIQUE INDEX uc_User_Connection_UserConnectionRank
 CREATE UNIQUE INDEX uc_User_Connection_UserConnectionId
    ON User_Connection (user_id ASC, provider_id ASC, provider_user_id ASC);
 
+DROP TABLE IF EXISTS User_Connection_Wechat;
+CREATE TABLE User_Connection_Wechat (
+    id              BIGINT          NOT NULL AUTO_INCREMENT,
+    user_id         BIGINT          NOT NULL,
+    app_id          VARCHAR(255)    NOT NULL,
+    open_id         VARCHAR(255)    NOT NULL,
+    union_id        VARCHAR(255)            ,
+    subscribe       TINYINT                 ,
+    session_key     VARCHAR(255)            ,
+    nickname        VARCHAR(255)            ,
+    sex             TINYINT                 ,
+    city            VARCHAR(255)            ,
+    province        VARCHAR(255)            ,
+    country         VARCHAR(255)            ,
+    language        VARCHAR(100)            ,
+    avatar_Url      VARCHAR(255)            ,
+    subscribe_time  BIGINT                  ,
+    remark          VARCHAR(255)            ,
+    PRIMARY KEY (id)
+);
+ALTER TABLE User_Connection_Wechat
+  ADD CONSTRAINT user_connection_wechat_user FOREIGN KEY (user_id)
+  REFERENCES `User` (id)
+  ON UPDATE NO ACTION
+  ON DELETE NO ACTION;    
+CREATE UNIQUE INDEX uc_User_Connection_Wechat_UserApp
+    ON User_Connection_Wechat (user_id ASC, app_id ASC);
+CREATE UNIQUE INDEX uc_User_Connection_Wechat_AppOpenId
+    ON User_Connection_Wechat (open_id ASC, app_id ASC);
+
+
 DROP TABLE IF EXISTS Authority;
 CREATE TABLE Authority (
     id              INT             NOT NULL,