+++ src/main/java/org/springframework/security/web/authentication/WechatMiniProperties.java
@@ -1,0 +1,14 @@
+package org.springframework.security.web.authentication;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import lombok.Data;
+
+@Data
+@ConfigurationProperties(prefix="wechat.mini")
+public class WechatMiniProperties {
+
+	private String appId;
+	private String secret;
+	
+}