+++ src/main/java/org/springframework/social/showcase/model/User.java
@@ -13,6 +13,7 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.JoinTable;
 import javax.persistence.ManyToMany;
+import javax.persistence.OneToMany;
 import javax.persistence.OneToOne;
 import javax.persistence.Transient;
 
@@ -20,8 +21,10 @@ import org.hibernate.annotations.GenericGenerator;
 import org.springframework.social.showcase.service.GenerateUniqueKey;
 
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 @Data
+@EqualsAndHashCode(callSuper=false, exclude="userConnectionWechatList")
 @Entity
 public class User {
 
@@ -43,6 +46,9 @@ public class User {
 	)
 	private Set<Authority> authorityList = new HashSet<Authority>();
 
+	@OneToMany(cascade=CascadeType.ALL, fetch=FetchType.LAZY, mappedBy="user")
+	private Set<UserConnectionWechat> userConnectionWechatList = new HashSet<UserConnectionWechat>();
+	
 	@Transient
 	public void assembleUser(String username, String password, String firstName, String lastName){
 		Account a = new Account(username, password, firstName, lastName);
@@ -54,6 +60,23 @@ public class User {
 	@Transient
 	public void addAuthority(Authority authority) {		
 		this.authorityList.add(authority);
+	}
+	
+	@Transient
+	public void addWechatConnection(UserConnectionWechat ucWechat) {
+		if(ucWechat!=null) {
+			ucWechat.setUser(this);			
+			this.userConnectionWechatList.add(ucWechat);
+		}
+	}	
+	@Transient 
+	public void createUserWithWechatConnection(String appId, String openId, String unionId, String sessionKey) {
+		UserConnectionWechat ucWechat = new UserConnectionWechat();
+		ucWechat.setAppId(appId);
+		ucWechat.setOpenId(openId);
+		ucWechat.setUnionId(unionId);
+		this.setHashId(GenerateUniqueKey.getInstance().generateUniqueKeyUsingMessageDigest());
+		this.addWechatConnection(ucWechat);
 	}
 	
 }